BitIO.c:
Power2Mask: Didn't work for 64 bit, only up to 31 bits. Fixed it so now
it works from 1-64 bits.

Removed the Bit-by-bit version of PeekBits with a byte-by-byte version,
because it passed all tests with flying colors, and now they fly up to
8 times faster.

Renamed NewPeekBits3 to PeekBits in the process.

Removed the old versions and ideas of PeekBits, cleaned up BitIO in
general.