Commented the header more extensively
Rearranged the code to logically group it.

Removed useless and irrelevant comments from the source.

I've still got some more testing and rewriting work to do in BitIO, and
of course to reintegrate the other files/headers like UUID, and MD5;
but it's getting pretty close to done.

Kinda sad, my first project has been a memorable one, I've learned and
grown SO much in the process of making it.