Tweaked the macros, and added Typeof independently of the compiler or mode (lookin' at you, Clang)
Also added an optimization to Insert and Extract Bits, so that when the
Bits2{Insert/Extract} is a multiple of 8, and so is the BitOFfset in
he BitBuffer, it’ll just copy bytes.

I did this for a nice speedup in byte orientated formats, like
BMP/PXM/WAV/AIF/W64 formats primarily, tho theoretically it’ll help
with some of the bit oriented formats at least sometimes.