ContainerIO: MAJOR changes to the AudioContainer and ImageContainer types.
There are now 2 channel masks per type, the first i just the channels present in the type, the second is an index into the image representing where each channel is located.

I've had to write a few new functions and some major changes internally to accept this change.

There were a few ABI and API changes, especially concerning the _Init functions, but overall, it's basically the same as it was before.

Overall, I was able to remove about 30 lines of code from ContainerIO as a result.